后端/requirements.txt
django==3.2.7
djangorestframework==3.12.4
django-rest-auth==0.9.5
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
idna==3.2
inflection==0.5.1
lxml==4.6.3
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
jmespath==0.10.0
django-allauth==0.45.0
urllib3==1.26.6
coverage==5.5
bs4==0.0.1
django-rest-swagger
